+++ packages/config/jest.config.js
@@ -1,9 +1,0 @@
-const config = require('../../jest/config');
-
-module.exports = Object.assign({}, config, {
-  coverageThreshold: {
-    global: {
-      lines: 90,
-    },
-  },
-});