+++ packages/loaders/test/plugin_loader_async.spec.ts
@@ -1,4 +1,5 @@
 import path from 'path';
+import { describe, expect, test } from 'vitest';
 
 import { Config, parseConfigFile } from '@verdaccio/config';
 import { pluginUtils } from '@verdaccio/core';