+++ packages/signature/test/legacy-token.spec.ts
@@ -1,3 +1,5 @@
+import { describe, expect, test } from 'vitest';
+
 import { aesDecrypt, aesEncrypt } from '../src';
 
 describe('test crypto utils', () => {