+++ packages/store/test/storage.spec.ts
@@ -44,7 +44,7 @@ function generateRandomStorage() {
   return path.join(tempRoot, tempStorage);
 }
 
-setup({ type: 'stdout', format: 'pretty', level: 'trace' });
+const logger = setup({ type: 'stdout', format: 'pretty', level: 'trace' });
 
 const domain = 'https://registry.npmjs.org';
 const fakeHost = 'localhost:4873';
@@ -149,7 +149,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -194,7 +194,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const ac = new AbortController();
         setTimeout(() => {
@@ -239,7 +239,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         // create a package
         const bodyNewManifest1 = generatePackageMetadata(pkgName, '1.0.0');
@@ -328,7 +328,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest1 = generatePackageMetadata(pkgName, '1.0.0');
         const bodyNewManifest2 = generatePackageMetadata(pkgName, '1.0.0');
@@ -365,7 +365,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
 
@@ -395,7 +395,7 @@ describe('storage', () => {
         const mockDate = '2018-01-14T11:17:40.712Z';
         MockDate.set(mockDate);
         const config = getConfig('deprecate.yaml');
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -442,7 +442,7 @@ describe('storage', () => {
         const mockDate = '2018-01-14T11:17:40.712Z';
         MockDate.set(mockDate);
         const config = getConfig('deprecate.yaml');
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         // publish new package
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
@@ -516,7 +516,7 @@ describe('storage', () => {
     describe('star', () => {
       test.each([['foo']])('star package %s', async (pkgName) => {
         const config = getConfig('deprecate.yaml');
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -549,7 +549,7 @@ describe('storage', () => {
         const mockDate = '2018-01-14T11:17:40.712Z';
         MockDate.set(mockDate);
         const config = getConfig('deprecate.yaml');
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -591,7 +591,7 @@ describe('storage', () => {
         const mockDate = '2018-01-14T11:17:40.712Z';
         MockDate.set(mockDate);
         const config = getConfig('deprecate.yaml');
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -630,7 +630,7 @@ describe('storage', () => {
 
       test.each([['foo']])('should unstar a package %s', async (pkgName) => {
         const config = getConfig('deprecate.yaml');
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -667,7 +667,7 @@ describe('storage', () => {
 
       test.each([['foo']])('should handle missing username %s', async (pkgName) => {
         const config = getConfig('deprecate.yaml');
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -695,7 +695,7 @@ describe('storage', () => {
         ['foo', 'publishWithOwnerAndCheck.yaml'],
       ])('new package %s, %s (anonymous)', async (pkgName, configFile) => {
         const config = getConfig(configFile);
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         await storage.updateManifest(bodyNewManifest, {
@@ -717,7 +717,7 @@ describe('storage', () => {
         ['foo', 'publishWithOwnerAndCheck.yaml'],
       ])('new package %s, %s (logged in)', async (pkgName, configFile) => {
         const config = getConfig(configFile);
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const owner = { name: 'fooUser', email: '' };
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
@@ -742,7 +742,7 @@ describe('storage', () => {
         ['foo', 'publishWithOwnerAndCheck.yaml'],
       ])('add/remove owner %s, %s', async (pkgName, configFile) => {
         const config = getConfig(configFile);
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const firstOwner = { name: 'fooUser', email: '' };
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
@@ -802,7 +802,7 @@ describe('storage', () => {
         ['foo', 'publishWithOwnerAndCheck.yaml'],
       ])('should fail removing last owner %s, %s', async (pkgName, configFile) => {
         const config = getConfig(configFile);
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
         const owner = 'fooUser';
@@ -830,7 +830,7 @@ describe('storage', () => {
         'ok to publish as non-owner without check %s, %s',
         async (pkgName, configFile) => {
           const config = getConfig(configFile);
-          const storage = new Storage(config);
+          const storage = new Storage(config, logger);
           await storage.init(config);
           const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
           const owner = 'fooUser';
@@ -860,7 +860,7 @@ describe('storage', () => {
         'should fail publishing as non-owner with check %s, %s',
         async (pkgName, configFile) => {
           const config = getConfig(configFile);
-          const storage = new Storage(config);
+          const storage = new Storage(config, logger);
           await storage.init(config);
           const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
           const owner = 'fooUser';
@@ -898,7 +898,7 @@ describe('storage', () => {
           storage: generateRandomStorage(),
         })
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       storage.init(config).then(() => {
         const ac = new AbortController();
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
@@ -938,7 +938,7 @@ describe('storage', () => {
           storage: generateRandomStorage(),
         })
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       storage.init(config).then(() => {
         const abort = new AbortController();
         storage
@@ -978,7 +978,7 @@ describe('storage', () => {
           __dirname
         )
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       storage.init(config).then(() => {
         const abort = new AbortController();
         storage
@@ -1022,7 +1022,7 @@ describe('storage', () => {
           __dirname
         )
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       storage.init(config).then(() => {
         const ac = new AbortController();
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
@@ -1086,7 +1086,7 @@ describe('storage', () => {
           __dirname
         )
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       storage.init(config).then(() => {
         const req = httpMocks.createRequest({
           method: 'GET',
@@ -1139,7 +1139,7 @@ describe('storage', () => {
           __dirname
         )
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       storage.init(config).then(() => {
         const ac = new AbortController();
         const bodyNewManifest = generatePackageMetadata(pkgName, '1.0.0');
@@ -1196,7 +1196,7 @@ describe('storage', () => {
           )
         );
 
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.syncUplinksMetadata(fooManifest.name, null, {
@@ -1220,7 +1220,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.syncUplinksMetadata(fooManifest.name, null, {
@@ -1241,7 +1241,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         const [manifest] = await storage.syncUplinksMetadata(fooManifest.name, fooManifest, {
           retry: { limit: 0 },
@@ -1263,7 +1263,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
 
         const [response] = await storage.syncUplinksMetadata(fooManifest.name, fooManifest);
@@ -1304,7 +1304,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
 
         const [response] = await storage.syncUplinksMetadata(fooManifest.name, null);
@@ -1326,7 +1326,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
 
         const [response] = await storage.syncUplinksMetadata(fooManifest.name, fooManifest);
@@ -1350,7 +1350,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
 
         const [response] = await storage.syncUplinksMetadata(fooManifest.name, fooManifest, {
@@ -1379,7 +1379,7 @@ describe('storage', () => {
             __dirname
           )
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
 
         const [response] = await storage.syncUplinksMetadata('foo', null, {
@@ -1400,7 +1400,7 @@ describe('storage', () => {
           storage: generateRandomStorage(),
         })
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       await storage.init(config);
       await expect(storage.getLocalDatabase()).resolves.toHaveLength(0);
     });
@@ -1420,7 +1420,7 @@ describe('storage', () => {
         },
         url: '/',
       });
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       await storage.init(config);
       const manifest = generatePackageMetadata('foo');
       const ac = new AbortController();
@@ -1449,7 +1449,7 @@ describe('storage', () => {
             storage: generateRandomStorage(),
           })
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await storage.saveToken({
           user: 'foo',
@@ -1471,7 +1471,7 @@ describe('storage', () => {
             storage: generateRandomStorage(),
           })
         );
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await storage.saveToken({
           user: 'foo',
@@ -1498,7 +1498,7 @@ describe('storage', () => {
           storage: generateRandomStorage(),
         })
       );
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       await storage.init(config);
       await expect(storage.removeTarball('foo', 'foo-1.0.0.tgz', 'rev', username)).rejects.toThrow(
         API_ERROR.NO_PACKAGE
@@ -1524,7 +1524,7 @@ describe('storage', () => {
         },
         url: '/',
       });
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       await storage.init(config);
 
       const manifest = generatePackageMetadata('foo');
@@ -1573,7 +1573,7 @@ describe('storage', () => {
 
     test('ok to remove package as non-owner without check', async () => {
       const config = getConfig('publishWithOwnerDefault.yaml');
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       await storage.init(config);
       const owner = 'fooUser';
       const options = { ...defaultRequestOptions, username: owner };
@@ -1612,7 +1612,7 @@ describe('storage', () => {
 
     test('should fail as non-owner with check', async () => {
       const config = getConfig('publishWithOwnerAndCheck.yaml');
-      const storage = new Storage(config);
+      const storage = new Storage(config, logger);
       await storage.init(config);
       const owner = 'fooUser';
       const options = { ...defaultRequestOptions, username: owner };
@@ -1661,7 +1661,7 @@ describe('storage', () => {
           },
           url: '/',
         });
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.getPackageByOptions({
@@ -1693,7 +1693,7 @@ describe('storage', () => {
           },
           url: '/',
         });
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.getPackageByOptions({
@@ -1726,7 +1726,7 @@ describe('storage', () => {
           },
           url: '/',
         });
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.getPackageByOptions({
@@ -1759,7 +1759,7 @@ describe('storage', () => {
           },
           url: '/',
         });
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.getPackageByOptions({
@@ -1799,7 +1799,7 @@ describe('storage', () => {
           },
           url: '/',
         });
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.getPackageByOptions({
@@ -1839,7 +1839,7 @@ describe('storage', () => {
           },
           url: '/',
         });
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
         await expect(
           storage.getPackageByOptions({
@@ -1873,7 +1873,7 @@ describe('storage', () => {
           },
           url: '/',
         });
-        const storage = new Storage(config);
+        const storage = new Storage(config, logger);
         await storage.init(config);
 
         const manifest = (await storage.getPackageByOptions({