+++ packages/proxy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @verdaccio/proxy
 
+## 7.0.0-next-7.20
+
+### Patch Changes
+
+- @verdaccio/core@7.0.0-next-7.20
+- @verdaccio/config@7.0.0-next-7.20
+- @verdaccio/utils@7.0.0-next-7.20
+
 ## 7.0.0-next-7.19
 
 ### Patch Changes